[cleanup] Cleanup code style and removed unneeded configurations
